HardwareObject name change since class is compatible with more MD versions
